This is first version of two pages. Index_page.html, css and main_page.html, css.
