+++ index.html
@@ -1,27 +1,25 @@
 <!DOCTYPE html>
-<html>
+<html lang="EN">
     <head>
-        <title>jobbar</title>
-        <link rel="stylesheet" href="/assets/indexStyle.css">
-        <link rel="stylesheet" href="/css/bootstrap.min.css">
+        <title>JOBBAR</title>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="stylesheet" type="text/css" href="/assets/index_page.css">
+        <link rel="stylesheet" href="/assets/index_page.css">
         <script src="/js/jquery.min.js"></script>
-        <script src="/js/bootstrap.min.js"></script>
+        <script type="text/javascript" src="/public/js/index.js"></script>
+        <link href="https://fonts.googleapis.com/css?family=Droid+Serif:400,700" rel="stylesheet">
+        <link rel="stylesheet" href="/assets/font-awesome-4.7.0/css/font-awesome.min.css">
     </head>
     <body>
-        <h1>Welcome to the WebApp</h1>
-        <div class="col-lg-6">
-            <div class="jumbotron">
-                <h1>Empoloyees</h1>
-                <p>Working with employess</p>
-                <p><a class="btn btn-primary btn-lg" href="/employees" role="button">Examin</a></p>
-            </div>
-        </div>
-        <div class="col-lg-6">
-            <div class="jumbotron">
-                <h1>Shifts</h1>
-                <p>Working with shifts</p>
-                <p><a class="btn btn-primary btn-lg" href="#" role="button">Examin</a></p>
-            </div>
-        </div>
+        <i class="fa fa-bars fa-3x" onmouseover="toggleMenu()" ></i>
+        <nav id="first_navigation" >
+            <ul id="menu-box">
+                <li><a href="/views/main.html">main</a></li>
+                <li><a href="about_us.html">about us</a></li>
+                <li><a href="folow_us.html">folow us</a></li>
+                <li><a href="Contact Us">contact us</a></li>
+            </ul>
+        </nav>
     </body>
 </html>
