--- public/assets/indexStyle.css
@@ -1,16 +1,0 @@
-body{
-    background: skyblue;
-    font-family: Verdana;
-    color: #fff;
-    padding: 30px;
-    }
-h1{
-    font-size: 48px;
-    text-transform: uppercase;
-    letter-spacing: 2px;
-    text-align: center;
-    }
-p{
-    font-size: 20px;
-    text-align: center;
-    }
