+++ src/main/webapp/public/assets/font-awesome-4.7.0/less/screen-reader.less
@@ -1,0 +1,5 @@
+// Screen Readers
+// -------------------------
+
+.sr-only { .sr-only(); }
+.sr-only-focusable { .sr-only-focusable(); }
