+++ src/main/webapp/public/assets/font-awesome-4.7.0/scss/_fixed-width.scss
@@ -1,0 +1,6 @@
+// Fixed Width Icons
+// -------------------------
+.#{$fa-css-prefix}-fw {
+  width: (18em / 14);
+  text-align: center;
+}
