+++ src/main/webapp/public/assets/index_page.css
@@ -1,0 +1,44 @@
+*{
+    margin: 0px;
+    padding: 0px;
+    box-sizing: border-box;
+    font-family: 'Droid Serif', serif;
+}
+body {
+    background: url('/images/2.jpg')no-repeat center center;
+    background-size: cover;
+    width: 100vh;
+    height: 100vh;
+}
+i{
+    padding: 6%;
+    color: rgb(255,255,255);
+    margin-left: 30%;
+}
+#menu-box{
+    position: absolute;
+    top: 0px;
+    width: 16%;
+    height: 100vh; 
+    background:rgba(3, 244, 237, .21);
+    border-right: 2px solid rgba(31, 35, 47,.49);
+    padding: 5%;
+    list-style-type: none;
+}
+nav ul li{
+    padding-top: 80%;
+}
+nav ul li a{
+    color:rgb(255,255,255);
+    text-decoration: none;
+    text-transform: uppercase;
+    font-size: 1.2em;
+    font-weight: bolder;
+}
+nav ul li a:hover{
+    color: rgba(31, 35, 47,.49);
+    text-shadow: 2px 2px 2px;
+}
+#menu-box{
+    display: none;
+}
