+++ src/main/webapp/views/main.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="EN">
+<link rel="stylesheet" type="text/css" href="/assets/main_page.css">
+    <body>
+        <div id="wrap-section">
+            <div class="wrap-botton">
+                <div class="botton">
+                    <h1>Empoloyees</h1>
+                    <p>Working with employess</p>
+                    <p><a class="" href="/employees" role="button">Examin</a></p>
+                </div>
+            </div>
+            <div class="wrap-botton">
+                <div class="botton second">
+                    <h1>Shifts</h1>
+                    <p>Working with shifts</p>
+                    <p><a class="" href="#" role="button">Examin</a></p>
+                </div>
+            </div>
+        </div>
+    </body>
+</html>
