--- views/newEmpl.ejs
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" href="/css/bootstrap.min.css">
-        <script src="/js/jquery.min.js"></script>
-        <script src="/js/bootstrap.min.js"></script>
-    </head>
-    <body>
-        <h1>Add new employee</h1>
-        <form method="POST">
-            <input type="text" name = "name" placeholder="Employee name..." required>
-            <input type="text" name = "surname" placeholder="Employee surname" required>
-            <input type="text" name = "position" placeholder="Employee position" required>
-            <input type="number" name = "sallary" placeholder="Employee sallary" required>
-            <input type="date" name = "birth" placeholder="Employee birth" required>
-                <button type="submit">Add new employee</button>
-        </form>
-    </body>
-</html>
